Change common dependencies strategy
Prior to this commit, the common dependencies were managed through
symlinks (which are supported in AzureML). This made it confusing
to many users and usage was fiddly.

The new method consists of isolating runs before submission so
both experiment code and common code makes it into the AzureML
job. This hopefully will make things more transparent.